pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "contextual-pii-tagger"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "faker>=25.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "accelerate>=0.30",
    "bitsandbytes>=0.43",
    "datasets>=2.19",
    "peft>=0.10",
    "pytest>=8.0",
    "scikit-learn>=1.4",
    "spacy>=3.7",
    "torch>=2.2",
    "transformers>=4.40",
    "trl>=0.8",
    "xgboost>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["test"]
